Try to be better about when to notify about single item moves.
In general, we don't want to notify about single item moves when we're swiping or dragging, since that messes up the animations. But, we would like to notify about them otherwise, and we currently don't do that. This commit is an attempt to better detect when and when not to notify about single item moves.